refactor(player): extract utility functions from videoplayer to index.ts
